Fix the issues

+++ src/CloudService/custom/Remove-AzCloudService.ps1
@@ -18,37 +18,12 @@
 Deletes a cloud service.
 .Description
 Deletes a cloud service.
-.Example
-PS C:\> {{ Add code here }}
-
-{{ Add output here }}
-.Example
-PS C:\> {{ Add code here }}
-
-{{ Add output here }}
-
-.Inputs
-Microsoft.Azure.PowerShell.Cmdlets.CloudService.Models.ICloudServiceIdentity
-.Outputs
-System.Boolean
-.Notes
-COMPLEX PARAMETER PROPERTIES
-
-To create the parameters described below, construct a hash table containing the appropriate properties. For information on hash tables, run Get-Help about_Hash_Tables.
-
-INPUTOBJECT <ICloudServiceIdentity>: Identity Parameter
-  [CloudServiceName <String>]: 
-  [Id <String>]: Resource identity path
-  [ResourceGroupName <String>]: 
-  [RoleInstanceName <String>]: Name of the role instance
-  [SubscriptionId <String>]: Subscription credentials which uniquely identify Microsoft Azure subscription. The subscription ID forms part of the URI for every service call.
-  [UpdateDomain <Int32?>]: Specifies the update domain
 .Link
 https://docs.microsoft.com/en-us/powershell/module/az.cloudservice/remove-azcloudservice
 #>
 function Remove-AzCloudService {
     [OutputType([System.Boolean])]
-    [CmdletBinding(DefaultParameterSetName='Delete', PositionalBinding=$false, SupportsShouldProcess, ConfirmImpact='Medium')]
+    [CmdletBinding(DefaultParameterSetName='Delete', PositionalBinding=$false, SupportsShouldProcess, ConfirmImpact='High')]
     param(
         [Parameter(ParameterSetName='Delete', Mandatory)]
         [Parameter(ParameterSetName='DeleteRoleInstance', Mandatory)]
@@ -75,18 +50,16 @@ function Remove-AzCloudService {
         ${SubscriptionId},
     
         [Parameter(ParameterSetName='DeleteViaIdentity', Mandatory, ValueFromPipeline)]
-        [Parameter(ParameterSetName='DeleteRoleInstanceViaIdentity', Mandatory, ValueFromPipeline)]
         [Microsoft.Azure.PowerShell.Cmdlets.CloudService.Category('Path')]
         [Microsoft.Azure.PowerShell.Cmdlets.CloudService.Models.ICloudServiceIdentity]
         # Identity Parameter
         # To construct, see NOTES section for INPUTOBJECT properties and create a hash table.
         ${InputObject},
 
-        [Parameter(ParameterSetName='DeleteRoleInstance')]
-        [Parameter(ParameterSetName='DeleteRoleInstanceViaIdentity')]
+        [Parameter(ParameterSetName='DeleteRoleInstance', Mandatory)]
         [Microsoft.Azure.PowerShell.Cmdlets.CloudService.Category('Body')]
-        [System.String[]]
-        # .
+        [System.String]
+        # Name of the role instance.
         ${RoleInstance},
     
         [Parameter()]
@@ -158,10 +131,16 @@ function Remove-AzCloudService {
     process {
         if ($PSBoundParameters.ContainsKey('RoleInstance')) {
             $PSBoundParameters.Add('CloudServiceName', $Name)
+            $PSBoundParameters.Add('RoleInstanceName', $RoleInstance)
             $Null = $PSBoundParameters.Remove('Name')
-            Az.CloudService.internal\Remove-AzCloudServiceRoleInstance @PSBoundParameters
+            $Null = $PSBoundParameters.Remove('RoleInstance')
+            if ($PSCmdlet.ShouldProcess("Will remove role instance $RoleInstance?")) {
+                Az.CloudService.internal\Remove-AzCloudServiceRoleInstance @PSBoundParameters
+            }
         } else {
-            Az.CloudService.internal\Remove-AzCloudService @PSBoundParameters
+            if ($PSCmdlet.ShouldProcess("Will remove cloud service?")) {
+                Az.CloudService.internal\Remove-AzCloudService @PSBoundParameters
+            }
         }
     }
 }
